added profile fields to users model class, added getters/setters, updated constructor; added profile fields to regDTO, updated users constructor in auth controller
